improvement: better handle expiration time
In case the key got renew, either by SET a new value or EXPIRE with new value, the entry still remains in the expiration queue. So when pop those entries out, we have to perform a stale-check, by making sure the key is actually expired.

+++ src/core/mod.rs
@@ -56,7 +56,9 @@ impl Core {
         match command {
             Command::ExpIntervalCheck => {
                 for key in self.storage.scan_expired_keys() {
-                    self.storage.delete(&key);
+                    if let Some(true) = self.storage.is_expire(&key) {
+                        self.storage.delete(&key);
+                    }
                 }
 
                 CommandResponse::Null
